+++ src/Pr0gramm/Provider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SocialiteProviders\Pr0Auth;
+namespace SocialiteProviders\Pr0gramm;
 
 use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\RequestOptions;
